Finalized implementation of fb_position() in the sense that the first call in outcheck.c produces the same output as seen from wgml 4.0 under the same conditions.
[git-p4: depot-paths = "//depot/openwatcom/": change = 34417]

+++ bld/wgml/h/copfiles.h
@@ -522,7 +522,7 @@ global bool             ps_device;      // true if device is PostScript
 global cop_device   *   bin_device;     // binary device being used
 global cop_driver   *   bin_driver;     // binary driver being used
 global cop_font     *   bin_fonts;      // binary fonts being used (linked list)
-global uint32_t         wgml_font_cnt;  // number of available fonts
+global uint16_t         wgml_font_cnt;  // number of available fonts
 global wgml_font    *   wgml_fonts;     // the available fonts
 
 /* Reset so can be reused with other headers. */
